# server-tspec/openapi.json
{
  "info": {
    "title": "xCustomServer API",
    "version": "1.0.0",
    "description": "A Xentral add-in under the heavenly might of the Little Dragon."
  },
  "openapi": "3.0.3",
  "paths": {
    "/tasks": {
      "get": {
        "operationId": "getTaskList",
        "tags": [
          "Tasks"
        ],
        "summary": "Get a list of tasks",
        "description": "Get all available tasks.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addTask",
        "tags": [
          "Tasks"
        ],
        "summary": "Create a task",
        "description": "Create a new task",
        "parameters": [],
        "requestBody": {
          "description": "Task schema Info",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "Task schema Info",
                "type": "object",
                "properties": {
                  "title": {
                    "description": "Task title",
                    "example": "Call John Doe",
                    "type": "string"
                  },
                  "description": {
                    "description": "Task description",
                    "example": "A note to call John Doe",
                    "type": "string"
                  },
                  "completed": {
                    "description": "Task completion status",
                    "example": false,
                    "type": "boolean"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "completed",
                  "description",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task schema Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "TasksApiSpec_delete_/tasks",
        "tags": [
          "Tasks"
        ],
        "summary": "jgjhgjh",
        "parameters": [],
        "responses": {}
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "getTask",
        "tags": [
          "Tasks"
        ],
        "summary": "Get a task",
        "description": "Get a specific task by ID.",
        "parameters": [
          {
            "description": "The identifier of the task to retrieve",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task schema Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Problem details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateTask",
        "tags": [
          "Tasks"
        ],
        "summary": "Update a task",
        "description": "Update a specific task by ID.",
        "parameters": [
          {
            "description": "The identifier of the task to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Task schema Info",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "Task schema Info",
                "type": "object",
                "properties": {
                  "title": {
                    "description": "Task title",
                    "example": "Call John Doe",
                    "type": "string"
                  },
                  "description": {
                    "description": "Task description",
                    "example": "A note to call John Doe",
                    "type": "string"
                  },
                  "completed": {
                    "description": "Task completion status",
                    "example": false,
                    "type": "boolean"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "completed",
                  "description",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task schema Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Problem details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "tags": [
          "Tasks"
        ],
        "summary": "Delete a task",
        "description": "Delete a specific task by ID.",
        "parameters": [
          {
            "description": "The identifier of the task to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "400": {
            "description": "Problem details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TaskResponse": {
        "description": "Task schema Info",
        "type": "object",
        "properties": {
          "id": {
            "description": "Task identifier",
            "example": 1,
            "type": "number"
          },
          "title": {
            "description": "Task title",
            "example": "Call John Doe",
            "type": "string"
          },
          "description": {
            "description": "Task description",
            "example": "A note to call John Doe",
            "type": "string"
          },
          "completed": {
            "description": "Task completion status",
            "example": false,
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "required": [
          "completed",
          "description",
          "id",
          "title"
        ]
      },
      "TaskRequest": {
        "description": "Task schema Info",
        "type": "object",
        "properties": {
          "title": {
            "description": "Task title",
            "example": "Call John Doe",
            "type": "string"
          },
          "description": {
            "description": "Task description",
            "example": "A note to call John Doe",
            "type": "string"
          },
          "completed": {
            "description": "Task completion status",
            "example": false,
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "required": [
          "completed",
          "description",
          "title"
        ]
      },
      "ProblemDetails": {
        "description": "Problem details",
        "type": "object",
        "properties": {
          "type": {
            "description": "A URI reference [RFC3986] that identifies the\nproblem type.  This specification encourages that, when\ndereferenced, it provide human-readable documentation for the\nproblem type (e.g., using HTML [W3C.REC-html5-20141028]).  When\nthis member is not present, its value is assumed to be\n\"about:blank\".",
            "type": "string"
          },
          "status": {
            "description": "The HTTP status code ([RFC7231], Section 6)\ngenerated by the origin server for this occurrence of the problem.",
            "type": "number"
          },
          "title": {
            "description": "A short, human-readable summary of the problem\ntype.  It SHOULD NOT change from occurrence to occurrence of the\nproblem, except for purposes of localization (e.g., using\nproactive content negotiation; see [RFC7231], Section 3.4).",
            "type": "string"
          },
          "invalid-params": {
            "description": "Invalid params list",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParams"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "status",
          "title",
          "type"
        ]
      },
      "InvalidParams": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Field name",
            "type": "string"
          },
          "reason": {
            "description": "Reason",
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "name",
          "reason"
        ]
      }
    }
  }
}